refactor: extract Message, ResourceSidebar, and CoffeeModal into components folder
